requirements.txt
azure-eventhub==1.3.1
environs
# arguments
argparse
# non-functional
pytest